I fixed the thing - Reduced the general-ledger/page from 1700 lines to 61. Refactored and improved loading speed with it
